feat: update react example
