+++ examples/cdn/payment.html
@@ -24,7 +24,7 @@
           href="form.html"
           class="py-1 px-4 border border-red-500 text-red-500 font-bold rounded hover:bg-red-500 hover:text-white transition"
         >
-          Create new payment
+          Create payment
         </a>
       </div>
     </div>
