+++ examples/react/src/main.tsx
@@ -1,9 +1,18 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import OrbitalWidget from './components/OrbitalWidget'
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom'
+import FormPage from './pages/FormPage'
+import PaymentPage from './pages/PaymentPage'
+import SuccessPage from './pages/SuccessPage'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <OrbitalWidget />
+    <Router>
+      <Routes>
+        <Route path="/" element={<FormPage />} />
+        <Route path="/payment" element={<PaymentPage />} />
+        <Route path="/success" element={<SuccessPage />} />
+      </Routes>
+    </Router>
   </StrictMode>
 )
