+++ examples/react/src/pages/PaymentPage.module.scss
